# pscanrules: Reduce Suspicious Comments rule JS FPs